Fix for CodeQL problem
There has been a change in CodeQL such that the
CMAKE_CXX_FLAGS_RELEASE macro doesn't exist when it first parses
CMakeLists.txt -- which results in an error in the REPLACE command
since it is missing the 4th argument. Placing the macro in quotes should
return an empty string which is a valid argument.

Adding the quote does not affect normal runs of cmake where the
macro is initialized, so the change is backwards compatible.