Let BUILD_SHARED_LIBS determine type of library
wxCLib is still a static library, but if BUILD_SHARED_LIBS=Yes then
a shared library is created instead of a static library. This allows the
developer to choose the type of library he or she wants.

Add an option to build a shared library

This first step is just to resolve the problems with getting a shared
library to work.