This isn't going to work, so removing lua completely
The build system for lua and wxLua is a lot more complicated than
I initially thought. It will work a lot better to either be in it's own
repository with wxSnapshot as a submodule, or be a folder
within the wxUiEditor repository where it will be primarily used.

For wxUiEditor, it can have a tight coupling with the wxSnapshot
submodule. For it's own repository, a dev will have a choice to
use wxSnapshot, or their own build of wxWidgets.